requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
requests>=2.25.0
google-generativeai>=0.3.0
pyyaml>=6.0
Pillow>=9.0.0
python-dotenv>=1.0.0

